Remove `Eq Term`, `Eq Type`, introduce `Aeq` newtype
Hashable states that for an instance to be lawful it should observe that
`a == b` implies `hash a == hash b`. Note that without this law, the
resulting hash cannot be used in probabilistic data structures such as
`HashMap`s. Because `Eq Term` and `Eq Type` are defined such that they
account for alpha equivalency, an instance for `Hashable Term` is
non-trivial - and indeed unable to be derived with `Generic`. Hence, the
instance removed in this commit did not observe the proper laws.

While this commit chooses to remove the instance, this isn't strictly
necessary:

 * Krzysztof Maziarz et al, have shown that a lawful instance is possible
   with reasonable complexity bounds[1]. However, its implementation is
   non-trivial, and Clash's only valid use of the instance is in its
   netlist stage, where it is easily replaced by a reliance on `Ord`.

 * `Eq` and `Ord` could be defined such that they do *not* account for
   alpha equivalency. But, this would make them harder to use in various
   containers.

This is the first commit in a series of two: this one removes `Eq Term`,
`Ord Term`, `Eq Type`, and `Ord Type`. This is done in order to
demonstrate that all the current uses of alpha equivalency comparisons
are converted to explicit alpha equivalency comparisons. This commit
will be followed by one that reintroduces `Eq` and `Ord`, whose
instances will not account for alpha equivalency.

[1] https://arxiv.org/abs/2105.02856