Use `doctest-parallel` for Prelude's doctests
From the project's README, this has a couple of advantages:

 * It runs tests in parallel
 * It runs tests against compiled code, instead of reinterpreting
   your whole project
 * It isolates examples in modules, ensuring your tests don't
   accidentally rely on each other
 * It parses cabal files to discover modules, no need for custom setup
   anymore!

The project lists two disadvantges, of which only one is relevant for
clash-compiler:

 * Use of conditionals in a cabal file as well as CPP flags will be
   ignored

This patch speeds up doctest from 183 seconds to 9 seconds on my
desktop, running a (12 core) 3900X. Note that this isn't a 1:1
comparison as this PR also moves some test to a module, avoiding some
pointless reevaluation.

Fixes #1622.